fix: Fixed db scheme

+++ schema.sql
@@ -7,7 +7,7 @@
 ***************************************************************/
 
 CREATE TABLE IF NOT EXISTS public.accounts (
-    id INT PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     username VARCHAR(255),
     balance DECIMAL(10, 2) DEFAULT 0.0
 );
@@ -16,7 +16,7 @@ CREATE TABLE IF NOT EXISTS public.users (
     USERNAME VARCHAR(128) NOT NULL PRIMARY KEY,
     PASSWORD TEXT NOT NULL,
     TOKEN TEXT,
-    account INT,
+    account BIGINT,
     CONSTRAINT fk_users_account FOREIGN KEY (account) REFERENCES public.accounts (id)
 );
